# openapi.yaml
openapi: 3.1.0
info:
  title: DataForGood API
  version: "1.0"
  description: |
    Route handlers de `app/api`. Las acciones del panel del SuperUsuario
    (`/sistema`, `/usuarios`, `/supervisar`) son server actions y no aparecen aquí.

    **Sesión de usuario:** se maneja con la cookie `session_token` (httpOnly). Ejecuta
    primero `POST /api/auth/login`; el navegador (o Insomnia) guarda la cookie
    y la manda sola en las siguientes peticiones.

    **SuperUsuario:** ejecuta `POST /api/auth/root/token` con la credencial de
    SuperUsuario del entorno (`ROOT_USER_ID` y su contraseña), copia `data.token`,
    pulsa **Authorize** y pégalo en `rootBearer`. Vale 2 horas. Con él responden
    `GET /api/usuarios` (todos los usuarios) y `GET /api/usuarios/{id}` (cualquier id).
    El resto de las rutas son de la cuenta de un usuario y piden `session_token`.

    **CSRF:** las mutaciones con un header `Origin` de otro sitio responden
    403 `Origen no permitido` (`proxy.ts`).

    Todas las respuestas JSON son `{ data }`, `{ message, data }` o `{ error }`.

    Este archivo se mantiene a mano: si cambias una ruta, actualízalo en el mismo commit.
# /api/docs/spec agrega al inicio el servidor desde el que se abre la página.
servers:
  - url: http://localhost:3000/dataforgood
    description: Docker local
  - url: http://localhost:3000
    description: next dev
  - url: https://multimodal-ai-lab.cicese.mx/dataforgood
    description: Producción
...
security:
  - sessionCookie: []
paths:
...
  /api/campanas:
    get:
      tags: [Campañas]
      summary: Listar campañas
      description: Usa uno de los filtros a la vez. Sin filtros lista las campañas visibles.
      security: []
      parameters:
        - { name: id, in: query, schema: { type: integer }, description: "Una sola campaña. La respuesta trae además `viewer: { isCreator, isMySupervision, baneado }` (`baneado`: el creador te baneó de esta campaña)" }
        - { name: available, in: query, schema: { type: boolean }, description: Campañas en las que puedo aportar (excluye aquellas de las que el creador me baneó) }
        - { name: mine, in: query, schema: { type: boolean }, description: Campañas que creé (requiere sesión) }
        - { name: supervised, in: query, schema: { type: boolean }, description: Campañas que superviso (requiere sesión) }
        - { name: misAportes, in: query, schema: { type: boolean }, description: "Campañas donde aporté, con `myContributionsCount`" }
      responses:
        "200":
          description: Campañas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Campaign" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Campañas]
      summary: Crear campaña
      description: |
        Exige sesión: la campaña es de quien la crea. `status` solo `borrador` (por defecto) o
        `en_revision`; con 5 campañas activas, solo `borrador`. Contadores, XP, `isSpecial` y
        `hasReviewerAssigned` del body se ignoran.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignInput" }
      responses:
        "201": { $ref: "#/components/responses/CampaignResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/datos/{id}/descarga:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Datos abiertos]
      summary: Descargar un conjunto de datos abiertos (ZIP)
      security: []
      responses:
        "200":
          description: Archivo ZIP
          content:
            application/zip:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
      description: Se obtiene con `POST /api/auth/login`. El navegador la manda sola.
    pendingVerificationCookie:
      type: apiKey
      in: cookie
      name: pending_verification_id
      description: La dejan el registro o un login sin verificar.
    rootBearer:
      type: http
      scheme: bearer
      description: Token del SuperUsuario. Se obtiene con `POST /api/auth/root/token` y vale 2 horas.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    BadRequest:
      description: Datos inválidos
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: No hay sesión (o la cuenta está suspendida o baneada)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: No existe
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    CampaignResult:
      description: Campaña resultante
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              data: { $ref: "#/components/schemas/Campaign" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    CampaignStatus:
      type: string
      enum: [borrador, en_revision, aceptada, activa, pausada, finalizada, rechazada]
    DataType:
      type: string
      enum: [texto, foto, video, audio, documento]
    CollectionMode:
      type: string
      enum: [checklist, texto_libre]
    ChecklistSeccion:
      type: object
      required: [titulo, opciones]
      properties:
        titulo: { type: string, maxLength: 120 }
        opciones: { type: array, minItems: 1, items: { type: string, maxLength: 120 } }
...
    CampaignInput:
      type: object
      required: [name, description, tematica, dataTypes, goalContributions, quotaPerUser, startDate, endDate, locationState, locationCity]
      properties:
        name: { type: string }
        description: { type: string }
        tematica: { type: string }
        status: { $ref: "#/components/schemas/CampaignStatus" }
        dataTypes: { type: array, minItems: 1, items: { $ref: "#/components/schemas/DataType" } }
        checklistSecciones:
          type: array
          description: >-
            Checklists con título que el participante marca además de la descripción (siempre obligatoria).
            Pueden ser cero o varios. Cada uno necesita título y al menos una opción; los títulos no se repiten.
            Hasta 50 checklists de 50 opciones, 120 caracteres por texto. Las respuestas del aporte se guardan como "Título: opción".
            Por compatibilidad, un `checklistOpciones` (lista plana) se guarda como un checklist titulado "Checklist".
          items: { $ref: "#/components/schemas/ChecklistSeccion" }
        goalContributions: { type: integer, minimum: 1 }
        quotaPerUser: { type: integer, minimum: 1 }
        startDate: { type: string, format: date }
        startTime: { type: string, example: "08:00" }
        endDate: { type: string, format: date }
        endTime: { type: string, example: "20:00" }
        locationState: { type: string }
        locationCity: { type: string }
        locationColonia: { type: string }
        organizer: { type: string }
      example:
        name: Baches en Tepic
        description: Fotos de baches en calles de la ciudad
        tematica: infraestructura
        status: en_revision
        dataTypes: [foto]
        checklistSecciones:
          - titulo: Tamaño del bache
            opciones: [Chico, Mediano, Grande]
          - titulo: Riesgo
            opciones: [Para autos, Para peatones]
        goalContributions: 100
        quotaPerUser: 5
        startDate: "2026-10-01"
        startTime: "08:00"
        endDate: "2026-10-31"
        endTime: "20:00"
        locationState: Nayarit
        locationCity: Tepic
        locationColonia: ""
        organizer: ""
    Campaign:
      type: object
      properties:
        id: { type: string }
        creatorId: { type: string }
        creatorName: { type: string }
        name: { type: string }
        description: { type: string }
        tag: { type: string }
        tematica: { type: string }
        status: { $ref: "#/components/schemas/CampaignStatus" }
        dataTypes: { type: array, items: { $ref: "#/components/schemas/DataType" } }
        collectionMode: { $ref: "#/components/schemas/CollectionMode", description: "Obsoleto: se deriva de checklistSecciones" }
        checklistOpciones: { type: array, items: { type: string }, description: "Obsoleto: usa checklistSecciones" }
        checklistSecciones: { type: array, items: { $ref: "#/components/schemas/ChecklistSeccion" } }
        goalContributions: { type: integer }
        quotaPerUser: { type: integer }
        currentContributions: { type: integer }
        approvedContributions: { type: integer }
        pendingContributions: { type: integer }
        rejectedContributions: { type: integer }
        participants: { type: integer }
        startDate: { type: [string, "null"], format: date }
        startTime: { type: [string, "null"] }
        endDate: { type: [string, "null"], format: date }
        endTime: { type: [string, "null"] }
        locationCity: { type: string }
        locationState: { type: string }
        locationColonia: { type: string }
        organizer: { type: string }
        xpPerContribution: { type: integer }
        isSpecial: { type: boolean }
        daysRemaining: { type: [integer, "null"] }
        hasReviewerAssigned: { type: boolean }
        isSaved: { type: boolean, description: Solo con sesión }
        isBanned: { type: boolean, description: "Solo con sesión: el creador te baneó de esta campaña" }
        myContributionsCount: { type: integer, description: Solo con `misAportes=true` }
...